+++ test/unit/lib/commands/Composed.mocks.js
@@ -8,7 +8,8 @@ const Mock = function () {
 
   const commandsStubs = {
     initUsers: sandbox.stub().usingPromise().resolves(),
-    dispatchAbilityAction: sandbox.stub().usingPromise().resolves()
+    dispatchAbilityAction: sandbox.stub().usingPromise().resolves(),
+    getAbilityState: sandbox.stub().usingPromise().resolves()
   }
 
   const stubs = {